working on part 1 of the project
adding the reading of files from argv[1] and error checking for file
input. Added loops to fork the process and then execute from that
moving forward.

TODO: check to ensure that more free's are not required to ensure that
there are no memory leaks.
TODO: add a makefile to compile each part of this project
individually.